dll to static lib

+++ susumu/src/platform/windows/windows_input.cpp
@@ -27,7 +27,7 @@ namespace susumu {
 		auto window = static_cast<GLFWwindow*>(App::Get().GetWindow().GetNativeWindow());
 		double xPos, yPos;
 		glfwGetCursorPos(window, &xPos, &yPos);
-		return { xPos, yPos };
+		return { (float)xPos, (float)yPos };
 	}
 
 	float WindowsInput::GetMouseXImpl()
